refactor: unify pool query in pool manager, deprecate x/gamm, remove from CL module
