+++ proto/osmosis/gamm/v1beta1/query.proto
@@ -36,8 +36,9 @@ service Query {
     option (google.api.http).get = "/osmosis/gamm/v1beta1/filtered_pools";
   }
 
-  // Per Pool gRPC Endpoints
+  // Deprecated: please use the alternative in x/poolmanager
   rpc Pool(QueryPoolRequest) returns (QueryPoolResponse) {
+    option deprecated = true;
     option (google.api.http).get = "/osmosis/gamm/v1beta1/pools/{pool_id}";
   }
 
@@ -106,10 +107,14 @@ service Query {
 }
 
 //=============================== Pool
+// Deprecated: please use the alternative in x/poolmanager
 message QueryPoolRequest {
+  option deprecated = true;
   uint64 pool_id = 1 [ (gogoproto.moretags) = "yaml:\"pool_id\"" ];
 }
+// Deprecated: please use the alternative in x/poolmanager
 message QueryPoolResponse {
+  option deprecated = true;
   google.protobuf.Any pool = 1 [ (cosmos_proto.accepts_interface) = "PoolI" ];
 }
