+++ proto/osmosis/poolmanager/v1beta1/query.yml
@@ -23,3 +23,8 @@ queries:
       query_func: "k.NumPools"
     cli:
       cmd: "NumPools"
+  Pool:
+    proto_wrapper:
+      query_func: "k.Pool"
+    cli:
+      cmd: "Pool"
