+++ x/poolmanager/client/query_proto_wrap.go
@@ -5,6 +5,8 @@ import (
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
+	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
+
 	"github.com/osmosis-labs/osmosis/v15/x/poolmanager"
 	"github.com/osmosis-labs/osmosis/v15/x/poolmanager/client/queryproto"
 	"github.com/osmosis-labs/osmosis/v15/x/poolmanager/types"
@@ -81,5 +83,23 @@ func (q Querier) EstimateSwapExactAmountOut(ctx sdk.Context, req queryproto.Esti
 func (q Querier) NumPools(ctx sdk.Context, _ queryproto.NumPoolsRequest) (*queryproto.NumPoolsResponse, error) {
 	return &queryproto.NumPoolsResponse{
 		NumPools: q.K.GetNextPoolId(ctx) - 1,
+	}, nil
+}
+
+// Pool returns the pool specified by id.
+func (q Querier) Pool(ctx sdk.Context, req queryproto.PoolRequest) (*queryproto.PoolResponse, error) {
+
+	pool, err := q.K.RoutePool(ctx, req.PoolId)
+	if err != nil {
+		return nil, status.Error(codes.Internal, err.Error())
+	}
+
+	any, err := codectypes.NewAnyWithValue(pool)
+	if err != nil {
+		return nil, err
+	}
+
+	return &queryproto.PoolResponse{
+		Pool: any,
 	}, nil
 }
